fix: remove unnecessary comments and move files to root
