+++ src/main/kotlin/com/example/taskManagerController/TaskManagerControllerApplication.kt
@@ -1,0 +1,11 @@
+package com.example.taskManagerController;
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+@SpringBootApplication
+class TaskManagerControllerApplication
+
+fun main(args: Array<String>) {
+    SpringApplication.run(TaskManagerControllerApplication::class.java, *args)
+}
