--- taskManagerController/src/main/kotlin/com/example/taskManagerController/TaskManagerControllerApplication.kt
@@ -1,11 +1,0 @@
-package com.example.taskManagerController;
-
-import org.springframework.boot.SpringApplication
-import org.springframework.boot.autoconfigure.SpringBootApplication
-
-@SpringBootApplication // 1
-class TaskManagerControllerApplication // 2
-
-fun main(args: Array<String>) { // 3
-    SpringApplication.run(TaskManagerControllerApplication::class.java, *args) // 4
-}
